picturas-border-tool-ms-main/pyproject.toml
[tool.poetry]
name = "picturas-border-tool-ms"
version = "0.1.0"
description = "PictuRAS microservice (tool) that applies a border to an input image, generating a new colored bordered output image."
authors = [
    "Afonso Costa <d13690@di.uminho.pt>",
    "Paulo Sousa <d7157@di.uminho.pt>"
]

[tool.poetry.dependencies]
python = "^3.9.13"
pika = "^1.3.2"
pydantic = "^2.10.3"
pillow = "^11.0.0"
boto3 = "^1.28.0"

[tool.poetry.dev-dependencies]
yapf = "^0.43.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
